# backend/docs/api-contract/openapi.json
{"openapi":"3.1.0","info":{"title":"医患信息数据服务云平台","description":"让不会写 SQL 的医护人员和病患，用大白话查到权威准确的医院数据；保证查询的中间过程不泄露患者信息。安全引擎：医盾（零 LLM 审计）。","version":"1.0.0"},"paths":{"/api/datasources":{"get":{"tags":["datasources"],"summary":"List Datasources","operationId":"list_datasources_api_datasources_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DatasourceInfo"},"type":"array","title":"Response List Datasources Api Datasources Get"}}}}}}},"/api/datasources/demo":{"post":{"tags":["datasources"],"summary":"Create Demo","description":"一键载入演示数据（建 5 表 + 灌虚构数据 + 初始化平台元数据库）。","operationId":"create_demo_api_datasources_demo_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/datasources/{datasource_id}/schema":{"get":{"tags":["datasources"],"summary":"Get Schema","operationId":"get_schema_api_datasources__datasource_id__schema_get","parameters":[{"name":"datasource_id","in":"path","required":true,"schema":{"type":"string","title":"Datasource Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/policies":{"get":{"tags":["policies"],"summary":"List Policies","description":"所有已定案策略的 id 列表（供数据源页展示）。","operationId":"list_policies_api_policies_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/policies/{datasource_id}":{"get":{"tags":["policies"],"summary":"Get Policy","operationId":"get_policy_api_policies__datasource_id__get","parameters":[{"name":"datasource_id","in":"path","required":true,"schema":{"type":"string","title":"Datasource Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["policies"],"summary":"Update Policy","description":"更新策略。只改 YAML 中明确提交的键，其余保留。","operationId":"update_policy_api_policies__datasource_id__put","parameters":[{"name":"datasource_id","in":"path","required":true,"schema":{"type":"string","title":"Datasource Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PolicyUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/policies/{datasource_id}/validate":{"get":{"tags":["policies"],"summary":"Validate Policy","description":"校验策略可被算法层正确加载（健康检查/接入新数据源时用）。","operationId":"validate_policy_api_policies__datasource_id__validate_get","parameters":[{"name":"datasource_id","in":"path","required":true,"schema":{"type":"string","title":"Datasource Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/query":{"post":{"tags":["query"],"summary":"Run Query","operationId":"run_query_api_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/query/direct":{"post":{"tags":["query"],"summary":"Run Direct Query","operationId":"run_direct_query_api_query_direct_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DirectQueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/metrics/detection":{"get":{"tags":["reports"],"summary":"Detection Metrics","operationId":"detection_metrics_api_metrics_detection_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/reports":{"get":{"tags":["reports"],"summary":"Reports","operationId":"reports_api_reports_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reports/{report_id}":{"get":{"tags":["reports"],"summary":"Report Detail","operationId":"report_detail_api_reports__report_id__get","parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reports/{report_id}/export":{"get":{"tags":["reports"],"summary":"Export Report","operationId":"export_report_api_reports__report_id__export_get","parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"integer","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/health":{"get":{"summary":"Health","description":"健康检查：算法层可用性 + 各路由挂载状态。","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AdmissionInfo":{"properties":{"passed":{"type":"boolean","title":"Passed"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"},"checked_tables":{"items":{"type":"string"},"type":"array","title":"Checked Tables"},"bound_to_subject":{"type":"boolean","title":"Bound To Subject","default":false}},"type":"object","required":["passed"],"title":"AdmissionInfo"},"DatasourceInfo":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"table_count":{"type":"integer","title":"Table Count"},"column_count":{"type":"integer","title":"Column Count"},"policy_ready":{"type":"boolean","title":"Policy Ready"}},"type":"object","required":["id","name","table_count","column_count","policy_ready"],"title":"DatasourceInfo"},"DegradationInfo":{"properties":{"level":{"type":"string","title":"Level"},"label":{"type":"string","title":"Label"},"message":{"type":"string","title":"Message","default":""}},"type":"object","required":["level","label"],"title":"DegradationInfo"},"DirectQueryRequest":{"properties":{"token":{"$ref":"#/components/schemas/Token"},"datasource_id":{"type":"string","title":"Datasource Id"},"question":{"type":"string","title":"Question"}},"type":"object","required":["token","datasource_id","question"],"title":"DirectQueryRequest","description":"端到端模式预留：自然语言直查（后续接入 NL→SQL 翻译时启用）。\n\n当前返回 501，前端契约已冻结此形状，避免将来契约变更。"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MetricsInfo":{"properties":{"elapsed_ms":{"type":"integer","title":"Elapsed Ms","default":0},"llm_calls":{"type":"integer","title":"Llm Calls","default":0},"db_access":{"type":"integer","title":"Db Access","default":0}},"type":"object","title":"MetricsInfo","description":"审计阶段的开销——**不是查询执行的开销**。\n\n`llm_calls` / `db_access` 恒为 0，这是「零 LLM、零查库」安全声明的\n可验证形式：审计器不调模型（无提示注入面）、不碰数据库（无困惑代理面）。\n查询本身的执行不计入本指标——`result` 非空即表示执行过。"},"PlanItem":{"properties":{"id":{"type":"integer","title":"Id"},"description":{"type":"string","title":"Description"},"sql_before":{"type":"string","title":"Sql Before"},"sql_after":{"type":"string","title":"Sql After"},"is_final":{"type":"boolean","title":"Is Final"}},"type":"object","required":["id","description","sql_before","sql_after","is_final"],"title":"PlanItem"},"PolicyUpdate":{"properties":{"column_labels":{"anyOf":[{"additionalProperties":{"additionalProperties":{"type":"string"},"type":"object"},"type":"object"},{"type":"null"}],"title":"Column Labels"},"cross_domain_rules":{"anyOf":[{"items":{"additionalProperties":true,"type":"object"},"type":"array"},{"type":"null"}],"title":"Cross Domain Rules"}},"type":"object","title":"PolicyUpdate"},"QueryRequest":{"properties":{"token":{"$ref":"#/components/schemas/Token"},"datasource_id":{"type":"string","title":"Datasource Id"},"question_id":{"type":"string","title":"Question Id"}},"type":"object","required":["token","datasource_id","question_id"],"title":"QueryRequest"},"QueryResponse":{"properties":{"admission":{"$ref":"#/components/schemas/AdmissionInfo"},"question":{"type":"string","title":"Question"},"plan":{"items":{"$ref":"#/components/schemas/PlanItem"},"type":"array","title":"Plan"},"events":{"items":{"$ref":"#/components/schemas/SecurityEvent"},"type":"array","title":"Events"},"rewrite":{"$ref":"#/components/schemas/RewriteInfo"},"degradation":{"$ref":"#/components/schemas/DegradationInfo"},"metrics":{"$ref":"#/components/schemas/MetricsInfo"},"result":{"anyOf":[{"$ref":"#/components/schemas/ResultSet"},{"type":"null"}]}},"type":"object","required":["admission","question","degradation"],"title":"QueryResponse"},"ResultSet":{"properties":{"columns":{"items":{"type":"string"},"type":"array","title":"Columns"},"rows":{"items":{"items":{},"type":"array"},"type":"array","title":"Rows"}},"type":"object","title":"ResultSet"},"RewriteInfo":{"properties":{"applied":{"type":"integer","title":"Applied","default":0},"log":{"items":{"type":"string"},"type":"array","title":"Log"}},"type":"object","title":"RewriteInfo"},"SecurityEvent":{"properties":{"sub_query_id":{"type":"string","title":"Sub Query Id"},"type":{"type":"string","title":"Type"},"type_label":{"type":"string","title":"Type Label"},"column":{"type":"string","title":"Column"},"severity":{"type":"string","title":"Severity"},"severity_label":{"type":"string","title":"Severity Label"},"detail":{"type":"string","title":"Detail"}},"type":"object","required":["sub_query_id","type","type_label","column","severity","severity_label","detail"],"title":"SecurityEvent"},"Token":{"properties":{"type":{"type":"string","enum":["staff","patient"],"title":"Type"},"subject_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject Id"}},"type":"object","required":["type"],"title":"Token"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}